[Refactor] Move resource URL generation to the schema
This is a more appropriate location for it, instead of the
resource attempting to work it out. As the resource is already
injected with the schema, it can now use its schema to generate
the URL.

As part of this change, the abstract schema class constructor
arguments have changed so that it is injected with the server it
belongs to, instead of the schema container. Child classes can
still access the schemas via the server instance.