[Feature] Support schemas for parent and interface model classes
Updates the schema container logic so that it can now resolve
schemas for model classes where a schema is registered for a
model's parent class or an interface implemented by the model.

Also adjusted the resource factory so that it relies more on the
schema resolution logic. This effectively means that the resource
factory can also create a JSON:API resource for a model based
on a parent class or implemented interface.