[Featute] Add helper method to correctly detect empty id
There are many places in the code base where we are using empty(),
but this is incorrect as an id of `"0"` should be allowed. This
commit adds a static method to the ResourceIdentifier class for
detecting whether an id is empty. It also fixes the ResourceObject
class so that it correctly accepts a non-empty id.